refactor: reduce complexity and fix test failures in InitiativeTracker
- Extract InitiativeTracker into focused components (CombatControls, InitiativeList, InitiativeCard)
- Reduce InitiativeTracker complexity from 13 to ~2-3
- Add performance optimizations with Map-based lookups and useMemo
- Extract helper functions from useInitiativeTracker hook to reduce complexity from 8 to ~3-4
- Add accessibility improvements with proper aria-labels and heading structure
- Fix all test failures in InitiativeTracker component tests (20/20 passing)
- Remove unused onEndCombat prop for cleaner interface
